+++ public/app/plugins/panel/logs/LogsPanel.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useMemo, useState, useEffect } from 'react';
+import React, { useCallback, useMemo, useLayoutEffect, useState, useEffect } from 'react';
 import { css } from '@emotion/css';
 import { LogRows, CustomScrollbar, LogLabels, useStyles2, usePanelContext } from '@grafana/ui';
 import {
@@ -17,7 +17,6 @@ import { getFieldLinksForExplore } from 'app/features/explore/utils/links';
 import { COMMON_LABELS } from '../../../core/logs_model';
 import { PanelDataErrorView } from 'app/features/panel/components/PanelDataErrorView';
 import { cloneDeep } from 'lodash';
-import useScrollTop from './useScrollTop';
 
 interface LogsPanelProps extends PanelProps<Options> {}
 
@@ -39,6 +38,7 @@ export const LogsPanel: React.FunctionComponent<LogsPanelProps> = ({
 }) => {
   const isAscending = sortOrder === LogsSortOrder.Ascending;
   const style = useStyles2(getStyles(title, isAscending));
+  const [scrollTop, setScrollTop] = useState(0);
   const [newData, setNewData] = useState(data);
 
   useEffect(() => {
@@ -72,7 +72,33 @@ export const LogsPanel: React.FunctionComponent<LogsPanelProps> = ({
     return [logRows, deduplicatedRows, commonLabels];
   }, [newData, dedupStrategy]);
 
-  const { scrollTop } = useScrollTop({ isAscending, logRows });
+  useLayoutEffect(() => {
+    const scrollbar = document.querySelector('.scrollbar-view');
+    // the height of each log
+    const logRowHeight = document.querySelector('tbody')?.children[0].clientHeight;
+    // the height of all of the logs
+    const logHeight = scrollbar?.clientHeight;
+    // the new height the log panel is going to have once the new log is added
+    const newLogHeight = logRowHeight && logHeight ? logRowHeight + logHeight : 0;
+
+    if (isAscending && scrollbar) {
+      if (newLogHeight < scrollbar.scrollHeight) {
+        // if the scrollbar is present, we check if the user is scrolled at the bottom to
+        // determine if we are gonna scroll all the way down when a new log comes in
+        // or if we are gonna keep our scroll position
+        if (scrollbar.scrollHeight - scrollbar.scrollTop === newLogHeight) {
+          setScrollTop(scrollbar.scrollHeight);
+        } else {
+          setScrollTop(scrollbar.scrollTop);
+        }
+      } else {
+        // if the scrollbar is not present, we scroll to the bottom by default
+        // so when it starts to be present, its at the bottom position by default
+        // until the user scrolls up
+        setScrollTop(scrollbar.scrollHeight);
+      }
+    }
+  }, [isAscending, logRows]);
 
   const getFieldLinks = useCallback(
     (field: Field, rowIndex: number) => {
